test(typechecker): split baseline support fixture
